🚧♻️ Moving and refactoring host dashboard files
