+++ app/Domits/src/features/hostdashboard/hostcalendar/styles/HostCalendarStyles.js
@@ -1,0 +1,210 @@
+import {StyleSheet} from "react-native";
+
+export const styles = StyleSheet.create({
+    body: {
+        flex: 1,
+        backgroundColor: "#f9f9f9",
+        paddingHorizontal: 16,
+        paddingVertical: 8,
+    },
+    spinnerContainer: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+        backgroundColor: "#ffffff",
+    },
+    noAccommodationsText: {
+        textAlign: "center",
+        fontSize: 18,
+        color: "#7f8c8d",
+        marginTop: 20,
+    },
+    calendarContent: {
+        flex: 1,
+        marginTop: 20,
+    },
+    header: {
+        fontSize: 24,
+        fontWeight: "bold",
+        textAlign: "center",
+        marginBottom: 16,
+        color: "#2c3e50",
+    },
+    pickerContainer: {
+        marginBottom: 20,
+        backgroundColor: "#fefefe",
+        borderRadius: 8,
+        borderWidth: 1,
+        borderColor: "#dcdcdc",
+        overflow: "hidden",
+    },
+    picker: {
+        height: 50,
+        width: "100%",
+        color: "#34495e",
+        marginBottom: 8,
+    },
+    calendarContainer: {
+        flex: 1,
+        marginTop: 10,
+        paddingHorizontal: 8,
+        paddingVertical: 16,
+        backgroundColor: "#ffffff",
+        borderRadius: 8,
+        borderWidth: 1,
+        borderColor: "#e0e0e0",
+        shadowColor: "#000",
+        shadowOffset: { width: 0, height: 2 },
+        shadowOpacity: 0.1,
+        shadowRadius: 4,
+    },
+    accommodationText: {
+        fontSize: 18,
+        fontWeight: "500",
+        marginBottom: 12,
+        textAlign: "center",
+        color: "#34495e",
+    },
+    alertMessage: {
+        textAlign: "center",
+        fontSize: 16,
+        fontStyle: "italic",
+        color: "#95a5a6",
+        marginTop: 20,
+    },
+    // Calendar Component
+    container: {
+        flex: 1,
+        padding: 16,
+        backgroundColor: '#f7f7f7',
+        color: 'black',
+    },
+    calendarComponentHeader: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        marginBottom: 16,
+        color: 'black',
+    },
+    monthYear: {
+        fontSize: 20,
+        fontWeight: 'bold',
+        color: 'black',
+    },
+    navButtons: {
+        flexDirection: 'row',
+    },
+    navButton: {
+        backgroundColor: 'transparent',
+        borderWidth: 0,
+        padding: 10,
+    },
+    navButtonText: {
+        fontSize: 20,
+        fontWeight: 'bold',
+        color: 'black',
+    },
+    datesContainer: {
+        flexDirection: 'row',
+        flexWrap: 'wrap',
+        justifyContent: 'flex-start',
+        padding: 10,
+        borderWidth: 1,
+        borderColor: '#e0e0e0',
+        borderRadius: 8,
+        backgroundColor: '#fafafa',
+    },
+    date: {
+        width: '14%',
+        height: 40,
+        justifyContent: 'center',
+        alignItems: 'center',
+        marginVertical: 3,
+        marginHorizontal: 4,
+        borderRadius: 8,
+        backgroundColor: '#ffffff',
+        borderWidth: 1,
+        borderColor: '#e0e0e0',
+        shadowColor: '#000',
+        shadowOffset: {width: 0, height: 2},
+        shadowOpacity: 0.1,
+        shadowRadius: 3,
+    },
+    activeDay: {
+        backgroundColor: '#e3f2fd',
+        borderColor: '#64b5f6',
+        shadowColor: '#64b5f6',
+        shadowOffset: {width: 0, height: 4},
+        shadowOpacity: 0.2,
+        shadowRadius: 4,
+    },
+    selectedDay: {
+        backgroundColor: '#c8e6c9',
+        borderColor: '#388e3c',
+        shadowColor: '#388e3c',
+        shadowOffset: {width: 0, height: 4},
+        shadowOpacity: 0.2,
+        shadowRadius: 4,
+        transform: [{scale: 1.05}],
+    },
+    startDay: {
+        backgroundColor: '#74c878',
+        borderColor: '#388e3c',
+        shadowColor: '#1b5e20',
+        shadowOffset: {width: 0, height: 4},
+        shadowOpacity: 0.3,
+        shadowRadius: 4,
+        transform: [{scale: 1.1}],
+    },
+    endDay: {
+        backgroundColor: '#74c878',
+        borderColor: '#388e3c',
+        shadowColor: '#1b5e20',
+        shadowOffset: {width: 0, height: 4},
+        shadowOpacity: 0.3,
+        shadowRadius: 4,
+        transform: [{scale: 1.1}],
+    },
+    range: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        padding: 8,
+        backgroundColor: '#ffffff',
+        borderRadius: 5,
+        marginVertical: 4,
+        borderWidth: 1,
+        borderColor: '#ddd',
+        color: 'black',
+    },
+    actions: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        marginTop: 16,
+        color: 'black',
+    },
+    button: {
+        paddingVertical: 10,
+        paddingHorizontal: 20,
+        borderRadius: 5,
+        justifyContent: 'center',
+        alignItems: 'center',
+        marginHorizontal: 8,
+        backgroundColor: '#4caf50',
+        color: 'white',
+    },
+    buttonText: {
+        fontSize: 16,
+        fontWeight: 'bold',
+        color: 'white',
+    },
+    UndoButton: {
+        backgroundColor: '#f44336',
+    },
+    SaveButton: {
+        backgroundColor: '#4caf50',
+    },
+    RemoveButton: {
+        backgroundColor: '#f44336',
+    },
+});
