+++ app/Domits/src/features/hostdashboard/screens/HostDashboardTab.js
@@ -9,10 +9,10 @@ import {
 } from 'react-native';
 import {SafeAreaView} from 'react-native-safe-area-context';
 import {Button} from '@aws-amplify/ui-react-native/src/primitives';
-import {useAuth} from '../../context/AuthContext';
+import {useAuth} from '../../../context/AuthContext';
 import {useNavigation} from '@react-navigation/native';
-import ImageSlider from '../utils/ImageSlider';
-import DateFormatterDD_MM_YYYY from '../DateFormatterDD_MM_YYYY';
+import ImageSlider from '../../../screens/utils/ImageSlider';
+import DateFormatterDD_MM_YYYY from "../../../screens/utils/DateFormatterDD_MM_YYYY";
 
 const HostDashboardTab = () => {
   const {userAttributes} = useAuth();
