+++ app/Domits/src/navigation/mainNavigationStack.js
@@ -19,20 +19,22 @@ import paymentMethods from "../screens/guestdashboard/GuestPaymentsTab";
 import reviews from "../screens/guestdashboard/GuestReviewsTab";
 import settings from "../screens/guestdashboard/GuestSettingsTab";
 import helpAndFeedback from "../screens/guestdashboard/GuestHelpAndFeedbackTab";
-import HostDashboard from "../screens/hostdashboard/HostDashboard";
-import HostCalendar from "../screens/hostdashboard/HostCalendarTab";
-import HostReviews from "../screens/hostdashboard/HostReviewsTab";
-import HostProfile from "../screens/hostdashboard/HostProfileTab";
+import HostDashboard from "../features/hostdashboard/screens/HostDashboard";
+import HostCalendar from "../features/hostdashboard/hostcalendar/screens/HostCalendarTab";
+import HostReviews from "../features/hostdashboard/hostreviews/screens/HostReviewsTab";
+import HostProfile from "../screens/hostprofile/HostProfileTab";
 import OnboardingHost from "../screens/apphostonboarding/OnboardingHost";
-import HostPayments from "../screens/hostdashboard/HostPaymentsTab";
-import HostListings from "../screens/hostdashboard/HostListingsTab";
-import HostSettings from "../screens/hostdashboard/HostSettingsTab";
-import emailSettings from "../screens/hostdashboard/emailSettings";
-import ListProperty from "../screens/hostdashboard/HostListingsTab";
+import HostPayments from "../features/hostdashboard/hostfinance/screens/HostPaymentsTab";
+import HostListings from "../features/hostdashboard/hostproperty/screens/HostListingsTab";
+import HostSettings from "../screens/hostprofile/HostSettingsTab";
+import emailSettings from "../screens/hostprofile/emailSettings";
+import HostHelpDesk from "../features/hostdashboard/hosthelpdesk/screens/HostHelpDesk";
+import ListProperty from "../features/hostdashboard/hostproperty/screens/HostListingsTab";
 import Register from "../screens/login/register";
 import ConfirmMail from "../screens/login/confirmMail";
-import HostDetailPage from "../screens/hostdashboard/HostDetailPage";
+import HostDetailPage from "../screens/propertyDetailsScreen/screens/HostDetailPage";
 import React from "react";
+import HostDashboardTab from "../features/hostdashboard/screens/HostDashboardTab";
 
 const Stack = createNativeStackNavigator();
 
@@ -60,16 +62,20 @@ function MainNavigationStack() {
         <Stack.Screen name="Reviews" component={reviews} />
         <Stack.Screen name="Settings" component={settings} />
         <Stack.Screen name="HelpAndFeedback" component={helpAndFeedback} />
-        {/*<Stack.Screen name="HostHomepage" component={HostHomepage} />*/}
+        {/* Host */}
+          {/*<Stack.Screen name="HostHomepage" component={HostHomepage} />*/}
         <Stack.Screen name="HostDashboard" component={HostDashboard} />
+        <Stack.Screen name="HostDashboardTab" component={HostDashboardTab} />
+        <Stack.Screen name="HostProfile" component={HostProfile} />
         <Stack.Screen name="HostCalendar" component={HostCalendar} />
         <Stack.Screen name="HostReviews" component={HostReviews} />
         <Stack.Screen name="OnboardingHost" component={OnboardingHost} />
         <Stack.Screen name="HostPayments" component={HostPayments} />
         <Stack.Screen name="HostListings" component={HostListings} />
         <Stack.Screen name="HostSettings" component={HostSettings} />
+        <Stack.Screen name="HostHelpDesk" component={HostHelpDesk} />
+
         <Stack.Screen name="emailSettings" component={emailSettings} />
-        {/*<Stack.Screen name="HostRevenue" component={HostRevenue} />*/}
         <Stack.Screen name="ListProperty" component={ListProperty} />
         {/*<Stack.Screen name="LocationFillIn" component={LocationFillIn} />*/}
         {/*<Stack.Screen name="selectAmenities" component={selectAmenities} />*/}
